fix: Fix UInt not supported for VssNodes
UInt could not be used before this commit because @JVMOverloads was
used for all generated VSS data classes. Since UInt is a Kotlin
specific data type it was incompatible with the annotation. As long as
we are only using one constructor param for VssSignals we can remove
the annotation because the default constructor will still be
available.

Note: The java byte code will still show an Int for the Kotlin UInt
model. This is fine as long as the SDK is speaking Kotlin while
interacting with these models.

close #93

+++ kuksa-sdk/src/main/kotlin/org/eclipse/kuksa/extension/vss/VssNodeCopyExtension.kt
@@ -92,8 +92,8 @@ fun <T : Any> VssSignal<T>.copy(datapoint: Datapoint): VssSignal<T> {
             BOOL -> bool
             INT32 -> int32
             INT64 -> int64
-            UINT32 -> uint32
-            UINT64 -> uint64
+            UINT32 -> uint32.toUInt()
+            UINT64 -> uint64.toULong()
             FLOAT -> float
             DOUBLE -> double
             STRING_ARRAY -> stringArray.valuesList
@@ -179,7 +179,7 @@ fun <T : VssNode> T.copy(
 // region Operators
 
 /**
- * Convenience operator for [deepCopy] with a [VssNode]. It will return the [VssNode] with the updated
+ * Convenience operator for [deepCopy] with a [VssNode]. It will return the parent [VssNode] with the updated child
  * [VssNode].
  *
  * @throws [IllegalArgumentException] if the copied types do not match.
